requirements.txt
# Core Python dependencies
numpy
matplotlib==3.7.2
pandas
scipy
pyyaml
tqdm
requests
setuptools
wheel

# ML + CV
torch
torchvision
torchaudio
opencv-python

# Extras from pip section
appdirs==1.4.4
click==8.1.8
docker-pycreds==0.4.0
easygui==0.98.3
einops==0.6.1
gitpython==3.1.44
imageio==2.37.0
meshio==3.0.0
pathtools==0.1.2
protobuf==4.25.7
psutil==7.0.0
sentry-sdk==2.27.0
setproctitle==1.3.6
smmap==5.0.2
solidspy==1.1.0.post1
wandb==0.15.2
